+++ src/components/dropdown-menu/dropdown-menu.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/nextjs";
-import DropdownMenu from "./DropdownMenu";
+import DropdownMenu from "./dropdown-menu";
 
 const meta = {
   title: "Components/DropdownMenu",
